+++ app/src/views/CountryPreparedness/PublicCountryPreparedness/index.tsx
@@ -1,11 +1,17 @@
 import {
     Fragment,
+    useCallback,
     useMemo,
 } from 'react';
+import { useParams } from 'react-router-dom';
+import { ArrowLeftLineIcon } from '@ifrc-go/icons';
 import {
+    BlockLoading,
+    Button,
     Container,
     Heading,
     Message,
+    TextOutput,
 } from '@ifrc-go/ui';
 import { useTranslation } from '@ifrc-go/ui/hooks';
 import { resolveToString } from '@ifrc-go/ui/utils';
@@ -15,38 +21,46 @@ import {
     isNotDefined,
 } from '@togglecorp/fujs';
 
-import useAuth from '#hooks/domain/useAuth';
+import Link from '#components/Link';
+import WikiLink from '#components/WikiLink';
+import useRouting from '#hooks/useRouting';
 import { useRequest } from '#utils/restRequest';
 
 import i18n from './i18n.json';
 import styles from './styles.module.css';
 
-interface Props {
-    perId?: number;
-}
-
-function PublicCountryPreparedness(props: Props) {
-    const {
-        perId,
-    } = props;
-    const { isAuthenticated } = useAuth();
+function PublicCountryPreparedness() {
     const strings = useTranslation(i18n);
+
+    const { perId, countryId } = useParams<{ perId: string, countryId: string }>();
+
+    const {
+        pending: perStatsResponsePending,
+        response: perStatsResponse,
+    } = useRequest({
+        skip: isNotDefined(countryId),
+        url: '/api/v2/public-per-stats/',
+        query: isDefined(countryId) && isDefined(perId) ? {
+            country: [Number(countryId)],
+            id: Number(perId),
+        } : undefined,
+    });
+
     const {
         pending: processStatusPending,
         response: processStatusResponse,
     } = useRequest({
-        skip: isNotDefined(perId) || isAuthenticated,
+        skip: isNotDefined(perId),
         url: '/api/v2/public-per-process-status/{id}/',
         pathVariables: {
             id: Number(perId),
         },
     });
-
     const {
         pending: assessmentResponsePending,
         response: assessmentResponse,
     } = useRequest({
-        skip: isNotDefined(processStatusResponse?.assessment) || isAuthenticated,
+        skip: isNotDefined(processStatusResponse?.assessment),
         url: '/api/v2/public-per-assessment/{id}/',
         pathVariables: {
             id: Number(processStatusResponse?.assessment),
@@ -57,12 +71,14 @@ function PublicCountryPreparedness(props: Props) {
         pending: prioritizationResponsePending,
         response: prioritizationResponse,
     } = useRequest({
-        skip: isNotDefined(processStatusResponse?.prioritization) || isAuthenticated,
+        skip: isNotDefined(processStatusResponse?.prioritization),
         url: '/api/v2/public-per-prioritization/{id}/',
         pathVariables: {
             id: Number(processStatusResponse?.prioritization),
         },
     });
+
+    const perStats = perStatsResponse?.results?.[0];
 
     const topFiveRatedComponents = useMemo(
         () => {
@@ -122,26 +138,85 @@ function PublicCountryPreparedness(props: Props) {
         [prioritizationResponse],
     );
 
-    const pending = processStatusPending
+    const { goBack } = useRouting();
+    const handleBackButtonClick = useCallback(() => {
+        goBack();
+    }, [goBack]);
+
+    const pending = perStatsResponsePending
+        || processStatusPending
         || assessmentResponsePending
         || prioritizationResponsePending;
-
-    if (pending) {
-        return (
-            <Message
-                className={styles.pendingMessage}
-                pending
-                description={strings.publicComponentFetchingData}
-            />
-        );
-    }
 
     return (
         <Container
             className={styles.publicCountryPreparedness}
             childrenContainerClassName={styles.preparednessContent}
+            heading={strings.nsPreparednessAndResponseCapacityHeading}
+            actionsContainerClassName={styles.actionsContainer}
+            headingLevel={2}
             withHeaderBorder
+            actions={(
+                <div className={styles.perAction}>
+                    <TextOutput
+                        label={strings.lastUpdatedLabel}
+                        value={processStatusResponse?.updated_at}
+                        valueType="date"
+                        strongValue
+                    />
+                    <WikiLink
+                        href="user_guide/Preparedness#how-to-use-it"
+                    />
+                </div>
+            )}
+            icons={(
+                <Button
+                    name={undefined}
+                    onClick={handleBackButtonClick}
+                    variant="tertiary"
+                    title={strings.goBackButtonTitle}
+                >
+                    <ArrowLeftLineIcon className={styles.backIcon} />
+                </Button>
+            )}
         >
+            <div className={styles.perOverviewDetails}>
+                <TextOutput
+                    label={strings.startDateLabel}
+                    value={perStats?.date_of_assessment}
+                    valueType="date"
+                    strongValue
+                />
+                <TextOutput
+                    label={strings.perPhaseLabel}
+                    value={processStatusResponse?.phase_display}
+                    strongValue
+                />
+                <TextOutput
+                    label={strings.focalPointNameLabel}
+                    value={perStats?.ns_focal_point_name}
+                    strongValue
+                />
+                <TextOutput
+                    label={strings.perCycleLabel}
+                    value={perStats?.assessment_number}
+                    valueType="number"
+                    strongValue
+                />
+                <TextOutput
+                    label={strings.typeOfAssessmentLabel}
+                    value={perStats?.type_of_assessment.name}
+                    strongValue
+                />
+                <TextOutput
+                    label={strings.focalPointEmailTitle}
+                    value={perStats?.ns_focal_point_email}
+                    strongValue
+                />
+            </div>
+            {pending && (
+                <BlockLoading className={styles.pendingMessage} />
+            )}
             {isDefined(topFiveRatedComponents) && (
                 <Container
                     heading={strings.publicHighlightedTopRatedComponentHeading}
@@ -187,6 +262,23 @@ function PublicCountryPreparedness(props: Props) {
                     )}
                 </Container>
             )}
+            <div className={styles.limitedAccess}>
+                <Message
+                    title={strings.componentLimitedAccess}
+                    description={strings.componentLimitedAccessDescription}
+                    actions={(
+                        <div className={styles.contactContainer}>
+                            <Link
+                                href="mailto:PER.Team@ifrc.org"
+                                external
+                                variant="secondary"
+                            >
+                                {strings.componentRequestSeeMore}
+                            </Link>
+                        </div>
+                    )}
+                />
+            </div>
         </Container>
     );
 }
